rename file name to match class name
