+++ src/CMakeLists.txt
@@ -30,7 +30,7 @@ add_library(RSA-Core
         utility.cpp
         key.cpp
         encryption.cpp
-        vec/operations.cpp
+        vec/base256.cpp
         vec/types.h
 )
 target_include_directories(RSA-Core PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
